requirements.txt
--index-url=https://insight:insightnexus@nexus.insightcapital.io/repository/pypi-all/simple
attrs==21.4.0
awscli==1.25.10
black==22.3.0
boto3==1.24.4
botocore==1.27.10
coverage==6.4.2
cryptography==36.0.2
Flask-Compress==1.11
Flask-Cors==3.0.10
Flask-SQLAlchemy==2.5.1
Flask==2.1.1
GitPython==3.1.27
honeycomb-beeline==3.2.0
insig==4.1.0
isort==5.10.1
jinja2-cli==0.8.2
jinja2==3.1.1
jsonschema==4.4.0
mock==4.0.3
moto==3.1.6
psycopg2-binary==2.9.3
pytest-cov==3.0.0
pytest-freezegun==0.4.2
pytest-mock==3.7.0
pytest==6.2.4
PyYAML==5.4.1
requests==2.27.1
sentry-sdk==1.5.10
ujson==5.2.0
